+++ app/Filament/Resources/Projects/Pages/EditProject.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\SAO\Filament\Resources\Projects\Pages;
+
+use Filament\Actions\DeleteAction;
+use Filament\Actions\ForceDeleteAction;
+use Filament\Actions\RestoreAction;
+use Filament\Resources\Pages\EditRecord;
+use Modules\SAO\Filament\Resources\Projects\ProjectResource;
+use Override;
+
+final class EditProject extends EditRecord
+{
+    #[Override]
+    protected static string $resource = ProjectResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            DeleteAction::make(),
+            ForceDeleteAction::make(),
+            RestoreAction::make(),
+        ];
+    }
+}