+++ app/Filament/Resources/Projects/ProjectResource.php
@@ -1,0 +1,76 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\SAO\Filament\Resources\Projects;
+
+use BackedEnum;
+use Coolsam\Modules\Resource;
+use Filament\Panel;
+use Filament\Schemas\Schema;
+use Filament\Support\Icons\Heroicon;
+use Filament\Tables\Table;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\SoftDeletingScope;
+use Modules\SAO\Filament\Resources\Projects\Pages\CreateProject;
+use Modules\SAO\Filament\Resources\Projects\Pages\EditProject;
+use Modules\SAO\Filament\Resources\Projects\Pages\ListProjects;
+use Modules\SAO\Filament\Resources\Projects\Schemas\ProjectForm;
+use Modules\SAO\Filament\Resources\Projects\Tables\ProjectsTable;
+use Modules\SAO\Models\Project;
+use Override;
+use UnitEnum;
+
+final class ProjectResource extends Resource
+{
+    protected static ?string $model = Project::class;
+
+    #[Override]
+    protected static string|UnitEnum|null $navigationGroup = 'SAO';
+
+    #[Override]
+    protected static ?int $navigationSort = 10;
+
+    protected static string|BackedEnum|null $navigationIcon = Heroicon::OutlinedRectangleStack;
+
+    protected static ?string $recordTitleAttribute = 'name';
+
+    public static function getSlug(?Panel $panel = null): string
+    {
+        return 'sao/projects';
+    }
+
+    public static function form(Schema $schema): Schema
+    {
+        return ProjectForm::configure($schema);
+    }
+
+    public static function table(Table $table): Table
+    {
+        return ProjectsTable::configure($table);
+    }
+
+    public static function getRelations(): array
+    {
+        return [
+            //
+        ];
+    }
+
+    public static function getPages(): array
+    {
+        return [
+            'index' => ListProjects::route('/'),
+            'create' => CreateProject::route('/create'),
+            'edit' => EditProject::route('/{record}/edit'),
+        ];
+    }
+
+    public static function getRecordRouteBindingEloquentQuery(): Builder
+    {
+        return parent::getRecordRouteBindingEloquentQuery()
+            ->withoutGlobalScopes([
+                SoftDeletingScope::class,
+            ]);
+    }
+}