+++ app/Filament/Resources/TicketTypes/Pages/EditTicketType.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\SAO\Filament\Resources\TicketTypes\Pages;
+
+use Filament\Actions\DeleteAction;
+use Filament\Actions\ForceDeleteAction;
+use Filament\Actions\RestoreAction;
+use Filament\Resources\Pages\EditRecord;
+use Modules\SAO\Filament\Resources\TicketTypes\TicketTypeResource;
+use Override;
+
+final class EditTicketType extends EditRecord
+{
+    #[Override]
+    protected static string $resource = TicketTypeResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            DeleteAction::make(),
+            ForceDeleteAction::make(),
+            RestoreAction::make(),
+        ];
+    }
+}